+++ fileio.h
@@ -1,7 +1,9 @@
+#include <stdint.h>
+
 typedef unsigned char byte;
 
-void MD5init(long *state);
-void MD5trans(long *state, byte *inpdata);
+void MD5init(int32_t *state);
+void MD5trans(int32_t *state, byte *inpdata);
 void MultRSA(byte *src, byte *dst, int numblocks);
 
 int ReadAsmFile(FILE *, byte *, int);
